fix(numpy): coerce np.float64 leakage points to native Python (1.2.2)
Three leakage paths where numpy array arithmetic produced np.float64
values that would corrupt VCP canonical hash if not pre-normalized:
- risk_models.py:72,74: daily_loss/weekly_loss from ndarray division
- strategies.py:311,765: float(x) * np.sign() produces np.float64
- audit.py: add _native() recursive sanitizer before canonical_json()

Root cause documented in docs/phase4_prep/NUMPY_2X_AUDIT.md.
Primary fix is in oprim 2.2.2 (canonical: repr(float(obj))).
These source-level float() casts are defense-in-depth.

+++ omodul/risk_models.py
@@ -69,9 +69,9 @@ def drawdown_circuit_breaker(
         )
 
     n = len(arr)
-    daily_loss = (arr[-1] - arr[-2]) / arr[-2]
+    daily_loss = float((arr[-1] - arr[-2]) / arr[-2])
     weekly_start_idx = max(0, n - 5)
-    weekly_loss = (arr[-1] - arr[weekly_start_idx]) / arr[weekly_start_idx]
+    weekly_loss = float((arr[-1] - arr[weekly_start_idx]) / arr[weekly_start_idx])
 
     dd_result = drawdown_curve(pd.Series(arr), input_type="equity")
     max_drawdown = float(dd_result["max_drawdown"])
